+++ main/Sensors.h
@@ -5,6 +5,7 @@
 #ifndef Sensors_h
 #define Sensors_h
 
+// @todo pin numbers
 #define PIN_SENSOR_TEMP 0
 #define PIN_SENSOR_HUMID 0
 #define PIN_SENSOR_SOIL 0
@@ -16,9 +17,9 @@ class Sensors
     Sensors();
     void begin();
 
-    int getTemperature();
-    int getHumidity();
-    int getSoilMoisture();
+    float getTemperature();
+    float getHumidity();
+    float getSoilMoisture();
     
   private:
     int lastTemperature;
